Added caching of query collections (and some work on receiving more adequate events for filtering query collections).
To be specific: let's say you do this:

    var coll = Task.all().filter("done", "=", true);
    coll.addEventListener('change', function() {
      alert("Collection changed!");
    });

And change the `done` property of a task object:

    myTask.done = true;

the `change` event will now be triggered (if the value of `done` was
`false` before), because `myTask` effectively has been added to the
collection. The same goes when `done` is set from true to false, because
it has now been removed from the collection.

Memory implications:
- Query collections are now cached, which is more efficient in some
  cases, but also mean they remain in memory after you start using them,
  including pointers to event subscribers.